==> code/5 label_0_1_su1.ipynb <==
import os
#import numpy as np
import pandas as pd

os.chdir('Z:/project')
train = pd.read_csv("./results/pat_visit_score_match_1202.csv")
train.shape

m1=0
n1=0
m2=0
n2=0
labelm=[]
labeln=[]
labelm2=[]
labeln2=[]
for index, row in train.iterrows():
    if row['m1']<10:
        m1=0
        labelm.append(m1)
    else:
        m1=1
        labelm.append(m1)
        
    if row['n1']<10:
        n1=0
        labeln.append(n1)
    else:
        n1=1
        labeln.append(n1)   
#    if row['m2']<3:
#        m2=0
#        labelm2.append(m2)
#    else:
#        m2=1
#        labelm2.append(m2)
        
#    if row['n2']<3:
#        n2=0
#        labeln2.append(n2)
#    else:
#        n2=1
#        labeln2.append(n2)
        
train['m1_1'] = labelm    
train['n1_1'] = labeln  
#train['m2_1'] = labelm2    
#train['n2_1'] = labeln2

train.to_csv("./results/train_2925_2.csv")

train.shape
